refactor: move dtos to controllers folder
