+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/UserController.java
@@ -1,19 +1,20 @@
 package br.db.ecotrack.ecotrack_api.controller;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserRequestDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserResponseDto;
-import br.db.ecotrack.ecotrack_api.service.CurrentUserService;
-import br.db.ecotrack.ecotrack_api.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.db.ecotrack.ecotrack_api.controller.request.UserRequestDto;
+import br.db.ecotrack.ecotrack_api.controller.response.UserResponseDto;
+import br.db.ecotrack.ecotrack_api.service.CurrentUserService;
+import br.db.ecotrack.ecotrack_api.service.UserService;
 import jakarta.persistence.EntityNotFoundException;
 
 @RestController
