+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
+package br.db.ecotrack.ecotrack_api.controller.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
