+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/request/MaterialRequestDto.java
@@ -1,0 +1,6 @@
+package br.db.ecotrack.ecotrack_api.controller.request;
+
+public record MaterialRequestDto(
+    String type,
+    String description) {
+}
