+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/request/UserRequestDto.java
@@ -1,0 +1,7 @@
+package br.db.ecotrack.ecotrack_api.controller.request;
+
+public record UserRequestDto(
+        String name,
+        String email,
+        String password) {
+}
