+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/response/LoginResponseDto.java
@@ -1,0 +1,8 @@
+package br.db.ecotrack.ecotrack_api.controller.response;
+
+public record LoginResponseDto(
+
+    String acessToken,
+    long expiresIn,
+    UserResponseDto user) {
+}
