+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/response/MaterialResponseDto.java
@@ -1,0 +1,7 @@
+package br.db.ecotrack.ecotrack_api.controller.response;
+
+public record MaterialResponseDto(
+    Long materialId,
+    String type,
+    String description) {
+}
