+++ src/main/java/br/db/ecotrack/ecotrack_api/controller/response/UserResponseDto.java
@@ -1,0 +1,7 @@
+package br.db.ecotrack.ecotrack_api.controller.response;
+
+public record UserResponseDto(
+    Long userId,
+    String name,
+    String email) {
+}
