--- src/main/java/br/db/ecotrack/ecotrack_api/domain/dto/LoginResponseDto.java
@@ -1,9 +1,0 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
-
-public record LoginResponseDto(
-
-    String acessToken,
-    long expiresIn,
-    UserResponseDto user
-    ) {
-}
