--- src/main/java/br/db/ecotrack/ecotrack_api/domain/dto/MaterialRequestDto.java
@@ -1,6 +1,0 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
-
-public record MaterialRequestDto(
-        String type,
-        String description) {
-}
