--- src/main/java/br/db/ecotrack/ecotrack_api/domain/dto/MaterialResponseDto.java
@@ -1,7 +1,0 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
-
-public record MaterialResponseDto (
-        Long materialId,
-        String type,
-        String description) {
-}
