--- src/main/java/br/db/ecotrack/ecotrack_api/domain/dto/UserRequestDto.java
@@ -1,7 +1,0 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
-
-public record UserRequestDto(
-    String name,
-    String email,
-    String password) {
-}
