--- src/main/java/br/db/ecotrack/ecotrack_api/domain/dto/UserResponseDto.java
@@ -1,7 +1,0 @@
-package br.db.ecotrack.ecotrack_api.domain.dto;
-
-public record UserResponseDto(
-        Long userId,
-        String name,
-        String email) {
-}
