+++ src/main/java/br/db/ecotrack/ecotrack_api/domain/mapper/MaterialMapper.java
@@ -2,15 +2,16 @@ package br.db.ecotrack.ecotrack_api.domain.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import br.db.ecotrack.ecotrack_api.domain.dto.MaterialRequestDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.MaterialResponseDto;
+
+import br.db.ecotrack.ecotrack_api.controller.request.MaterialRequestDto;
+import br.db.ecotrack.ecotrack_api.controller.response.MaterialResponseDto;
 import br.db.ecotrack.ecotrack_api.domain.entity.Material;
 
 @Mapper(componentModel = "spring")
 public interface MaterialMapper {
 
-    MaterialResponseDto toDto(Material material);
+  MaterialResponseDto toDto(Material material);
 
-    @Mapping(target = "materialId", ignore = true)
-    Material toEntity(MaterialRequestDto materialRequestDto);
+  @Mapping(target = "materialId", ignore = true)
+  Material toEntity(MaterialRequestDto materialRequestDto);
 }
