+++ src/main/java/br/db/ecotrack/ecotrack_api/domain/mapper/UserMapper.java
@@ -2,15 +2,16 @@ package br.db.ecotrack.ecotrack_api.domain.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserRequestDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserResponseDto;
+
+import br.db.ecotrack.ecotrack_api.controller.request.UserRequestDto;
+import br.db.ecotrack.ecotrack_api.controller.response.UserResponseDto;
 import br.db.ecotrack.ecotrack_api.domain.entity.User;
 
 @Mapper(componentModel = "spring")
 public interface UserMapper {
 
-    UserResponseDto toDto(User user);
+  UserResponseDto toDto(User user);
 
-    @Mapping(target = "userId", ignore = true)
-    User toEntity(UserRequestDto userRequestDto);
+  @Mapping(target = "userId", ignore = true)
+  User toEntity(UserRequestDto userRequestDto);
 }
