+++ src/main/java/br/db/ecotrack/ecotrack_api/service/AuthService.java
@@ -8,9 +8,9 @@ import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.jwt.JwtEncoderParameters;
 import org.springframework.stereotype.Service;
 
-import br.db.ecotrack.ecotrack_api.domain.dto.LoginRequestDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.LoginResponseDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserResponseDto;
+import br.db.ecotrack.ecotrack_api.controller.request.LoginRequestDto;
+import br.db.ecotrack.ecotrack_api.controller.response.LoginResponseDto;
+import br.db.ecotrack.ecotrack_api.controller.response.UserResponseDto;
 import br.db.ecotrack.ecotrack_api.domain.entity.User;
 import br.db.ecotrack.ecotrack_api.domain.mapper.UserMapper;
 import br.db.ecotrack.ecotrack_api.repository.UserRepository;
@@ -19,48 +19,48 @@ import jakarta.persistence.EntityNotFoundException;
 @Service
 public class AuthService {
 
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtEncoder jwtEncoder;
-    private final UserMapper userMapper;
+  private final UserRepository userRepository;
+  private final PasswordEncoder passwordEncoder;
+  private final JwtEncoder jwtEncoder;
+  private final UserMapper userMapper;
 
-    public AuthService(UserRepository userRepository, PasswordEncoder passwordEncoder, JwtEncoder jwtEncoder,
-            UserMapper userMapper) {
-        this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-        this.jwtEncoder = jwtEncoder;
-        this.userMapper = userMapper;
+  public AuthService(UserRepository userRepository, PasswordEncoder passwordEncoder, JwtEncoder jwtEncoder,
+      UserMapper userMapper) {
+    this.userRepository = userRepository;
+    this.passwordEncoder = passwordEncoder;
+    this.jwtEncoder = jwtEncoder;
+    this.userMapper = userMapper;
+  }
+
+  public LoginResponseDto login(LoginRequestDto loginRequestDto) {
+
+    User user = userRepository.findByEmail(loginRequestDto.email())
+        .orElseThrow(() -> new EntityNotFoundException("Email não encontrado:" + loginRequestDto.email()));
+
+    boolean passwordMatches = checkPassword(loginRequestDto.password(), user.getPassword());
+
+    if (!passwordMatches) {
+      throw new IllegalArgumentException("Senha incorreta");
     }
 
-    public LoginResponseDto login(LoginRequestDto loginRequestDto) {
+    long expiresIn = 600L;
 
-        User user = userRepository.findByEmail(loginRequestDto.email())
-                .orElseThrow(() -> new EntityNotFoundException("Email não encontrado:" + loginRequestDto.email()));
+    JwtClaimsSet jwt = JwtClaimsSet.builder()
+        .issuer("ecotrack-api")
+        .subject(user.getName())
+        .expiresAt(Instant.now().plusSeconds(expiresIn))
+        .issuedAt(Instant.now())
+        .claim("email", user.getEmail())
+        .build();
 
-        boolean passwordMatches = checkPassword(loginRequestDto.password(), user.getPassword());
+    String token = jwtEncoder.encode(JwtEncoderParameters.from(jwt)).getTokenValue();
+    UserResponseDto userResponseDto = userMapper.toDto(user);
 
-        if (!passwordMatches) {
-            throw new IllegalArgumentException("Senha incorreta");
-        }
+    return new LoginResponseDto(token, expiresIn, userResponseDto);
+  }
 
-        long expiresIn = 600L;
-
-        JwtClaimsSet jwt = JwtClaimsSet.builder()
-                .issuer("ecotrack-api")
-                .subject(user.getName())
-                .expiresAt(Instant.now().plusSeconds(expiresIn))
-                .issuedAt(Instant.now())
-                .claim("email", user.getEmail())
-                .build();
-
-        String token = jwtEncoder.encode(JwtEncoderParameters.from(jwt)).getTokenValue();
-        UserResponseDto userResponseDto = userMapper.toDto(user);
-
-        return new LoginResponseDto(token, expiresIn, userResponseDto);
-    }
-
-    private boolean checkPassword(String rawPassword, String encodedPassword) {
-        return passwordEncoder.matches(rawPassword, encodedPassword);
-    }
+  private boolean checkPassword(String rawPassword, String encodedPassword) {
+    return passwordEncoder.matches(rawPassword, encodedPassword);
+  }
 
 }
