+++ src/main/java/br/db/ecotrack/ecotrack_api/service/UserService.java
@@ -3,8 +3,9 @@ package br.db.ecotrack.ecotrack_api.service;
 import java.util.Optional;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserRequestDto;
-import br.db.ecotrack.ecotrack_api.domain.dto.UserResponseDto;
+
+import br.db.ecotrack.ecotrack_api.controller.request.UserRequestDto;
+import br.db.ecotrack.ecotrack_api.controller.response.UserResponseDto;
 import br.db.ecotrack.ecotrack_api.domain.entity.User;
 import br.db.ecotrack.ecotrack_api.domain.mapper.UserMapper;
 import br.db.ecotrack.ecotrack_api.repository.UserRepository;
